Compiler changes for better fine-grained unit tests
